+++ src/main/java/boblovespi/factoryautomation/common/tileentity/TEBlastFurnaceController.java
@@ -232,8 +232,6 @@ public class TEBlastFurnaceController extends BlockEntity
 	@Override
 	public ClientboundBlockEntityDataPacket getUpdatePacket()
 	{
-		CompoundTag nbt = new CompoundTag();
-		save(nbt);
 		return ClientboundBlockEntityDataPacket.create(this, BlockEntity::saveWithFullMetadata);
 	}